requirements.txt
# Render deployment - web voice agent only
# Excludes: livekit-agents, chromadb, sentence-transformers, sounddevice, streamlit, pyngrok
# This keeps the build under 3 minutes on Render free tier

flask>=3.0.0
gunicorn>=21.0.0
groq>=0.9.0
python-dotenv>=1.0.0
soundfile>=0.12.1
numpy>=1.24.0
pydub>=0.25.0
gtts>=2.4.0
edge-tts>=6.1.0
requests>=2.28.0
beautifulsoup4>=4.12.0
livekit-agents
livekit-plugins-groq
livekit-plugins-google
livekit-plugins-silero
langchain
langchain-openai
chromadb
pysqlite3-binary
